Rename policies files for integration tests
