+++ ui/apps/platform/cypress/constants/PoliciesPage.js
