+++ ui/apps/platform/cypress/integration/policies/importPolicy.test.js
@@ -1,7 +1,7 @@
 import * as api from '../../constants/apiEndpoints';
-import { selectors } from '../../constants/PoliciesPagePatternFly';
+import { selectors } from '../../constants/PoliciesPage';
 import withAuth from '../../helpers/basicAuth';
-import { doPolicyRowAction, visitPolicies } from '../../helpers/policiesPatternFly';
+import { doPolicyRowAction, visitPolicies } from '../../helpers/policies';
 
 describe('Import policy', () => {
     withAuth();
